Fix m5term select issues
Running m5term on Ubuntu (Linux on Windows) fails with no STDIN input.
This is due to several unexpected sanity checks in readwrite() polling
loop.

Returning 0 from select() means timeout which should be expected as there
is a 1 second timeout set to the select().
And it is actually no need to perform "read(nfd) < 0" sanity checks.

This patch fixes the select() loop related issues.

Signed-off-by: Lv Zheng <zhenglv@smart-core.cn>