mem-ruby: Move getNumControllers out of machines
Move the getNumControllers function out of the protocol-specific
machine definitions (which are now in their own namespaces) to a
Base<Machine>Machine class that the protocol-specific machines inherit
from. This design was the simplest way to keep backwards compatibility.

The one caveat is that this design means that while you can compile
multiple protocols into one binary, you cannot use the same machine type
from different protocols in one simulation. This is the current
assumption, so it shouldn't change anything for users.

Change-Id: If9eeb78c24eb5076956b9f44fa04c7e6eb0f0dd4
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>