mem-ruby,scons: Update scons to use split SLICC
This updates and refactors the SConscript file in ruby/protocol to use
the new split shared/protocol-specific mechanism and also refactors the
Sconscript to make it a bit cleaner.

One challenge: This changeset requires SLICC to be run 4 times instead
of just twice. For both the "shared" files and the protocol specific
files we have to run SLICC twice. In practice, even on a network file
system it doesn't add more than a few seconds to the scons startup time.

Change-Id: Id86d15377b0a3cfe040be080a881f7bbb9b7450d
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>