tests: Replace multi-isa test GCN3 compilation with VEGA
As of this commit,
https://gem5-review.googlesource.com/c/public/gem5/+/59012, all GCN3
tests were replaced with VEGA. However, this commit,
https://gem5-review.googlesource.com/c/public/gem5/+/57510, was added
and used GCN3_X86 instead of Vega.

It is likely this is the root cause of recent Kokoro timeouts as
compiling both GCN3_X86 and VEGA_X86 is costly. This patch fixes this.

Change-Id: Iee2a08189a73283877f5d599191e9d1b105322ca
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/60169
Maintainer: Bobby Bruce <bbruce@ucdavis.edu>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>