tests: Remove '--ignore-style' from m5 weekly test build
The '--ignore-style' flag was added to the `scons build/x86/out/m5`
command in the weekly tests in this commit:
https://gem5-review.googlesource.com/c/public/gem5/+/63012

The m5 compilation does not attempt to run the style check hooks and, as
such, this command failed as the '--ignore-style' flag is not
recognized.

This caused the weekly tests to fail:
https://jenkins.gem5.org/job/weekly/76

This commit removes this flag.

Change-Id: Ic0320609ac5234be978743377f13dd1cf7f1e782
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/63553
Reviewed-by: Bobby Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby Bruce <bbruce@ucdavis.edu>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>