tests: Fix errors in arm switcheroo
When switching to the all target, there is no longer a "DerivO3CPU" we
now need to drop the "Deriv"

Note that there are still some cases where it is valid to use "DerivO3"
where it is converted to drop the "Deriv" via a map.

Cleaning this up further is for future work.

Change-Id: Ide59eb3a5c7daf15c13afe395409f588562c1b54
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/63851
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Maintainer: Giacomo Travaglini <giacomo.travaglini@arm.com>
Reviewed-by: Giacomo Travaglini <giacomo.travaglini@arm.com>