sim: Add a way to specify the pools allocators for each process
The system supports the specification of multiple blocks of memory and the ability to specify different pool allocators. However, it seems that the relevant memory allocation functions currently use the default pool_id parameter, which prevents subsequent memory blocks from being used. This patch adds a parameter and modifies the allocation function to make it easier to specify the set of pool ids available for each process.