util,misc: Replace MAINTAINERS.yaml with TAGS.yaml
Since migrating to GitHub from Gerrit, our maintainers are defined in
the gem5 GitHub "maintainers" Team:
https://github.com/orgs/gem5/teams/maintainers.

We are no longer assigning project maintainers to specific subsystems
and therefore this mapping is no longer required.

TAGS.yaml is included here to give a definitive list of tags to put in
git commit message headers.

The pre-commit check code has been udpated and refactored to reflect
this change.

The CODE-OF-CONDUCT.md and CONTRIBUTING.md documents still reference
MAINTAINERS.yaml but these references will be removed when updating both
these documents to reflect the GitHub migration.

Change-Id: I9ff9daffc602bd814301bb63740e33920af17cb3