misc: Remove 'Change-Id' req check from CI Tests
As we were moving from a 'Squash and Merge' model to a 'Merge' model we
can no longer guarantee each commit has a Change-Id. Merge commits will
always be without a Change-Id and therefore cause the CI tests to break
on every Pull Request.

This change removes this requirement from the CI Tests. From this change
onward, having a Change-Id in each commit message is no longer a
requirement.

Change-Id: I14ec52a92da8be674fca5436702f7a1162647143