misc: Remove 'continue-on-error' from workload files
'continue-on-error' does allow for a job to run even if other jobs have
failed. However, if the job fails the 'continue-on-error' setting does
not set the workload as Failed if the job failed.
'if: ${{ success() || failure() }}' if the correct. This ensures a job
runs even if the Workload is in a pass or success state.

Change-Id: Ib7a98c8bcc24a3134b1058a1564bf6155edf9c3f