base: Add test to reproduce Issue #240
This test, `AddrRangeTest_isSubsetLastByte` checks that if an address
range , "A", ending in the last byte of a 64 bit space is not a subset
of an address range, "B",  starting at the beginning of the space.
https://github.com/gem5/gem5/issues/240 highlights that this does not
hold due to an integer overflow error where end of range "A" is set to
the beggining of the range, where "B" begins, thus making "A" a subset
of "B".

The AddressRange tests can be ran from the root of the gem5 repo with:

```sh
scons build/NULL/base/addr_range.test.opt
./build/NULL/base/addr_range.test.opt
```

Change-Id: I1636caf2991aa6ea07e421b24e7843dba7666548