arch-riscv: Change to VPUStatus::DIRTY for every rvv arith inst
This is similar to [1] and [2].

Essentially, the VS bits of STATUS CSR keep track of the state of
the vector registers. (VS bits == DIRTY) means the content of vector
registers have been updated since the last time the VS bits were updated.

This change is supposed to change the VS bits to DIRTY for if any
vector arithmetic instruction is executed.

[1] https://gem5-review.googlesource.com/c/public/gem5/+/65272
[2] https://github.com/gem5/gem5/pull/370

Change-Id: I0427890dadc63b74a470d7405807dcfcad18005b
Signed-off-by: Hoa Nguyen <hn@hnpl.org>