python,misc: Run `pre-commit autoupdate` and run all
`pre-commit autoupdate` updates all the hooks to the latest version.

`pre-commit run --all-files` then updates all the files in the repo to
be in compliance with the new hook standards.

in this case the vast majority of updates were due to Python Black
enforcing stricter rules on new lines characters.

Change-Id: Ied4a21e7d099e3fe7ef392d1c2f01fefee5a0462