util: Create resource api and DummyResource stub
Here I've created a `AbstractDataSource` class. I use this to define the
API i need an expect for _any_ gem5 resources datasource I'm going to
connect to. With that defined I create a stub `DummyDatabase` to
imlement this and check it's functionality. There's no need at this
stage to worry about database interaction, or even writing out to file.

The layout, structure, and design methodology is what i'm trying to get
across. The API is likely too basic for our needs and nothing is
complete (I didn't even test it), but it should give an idea of how
things can be laid out in a cleaner, more managable way.

Change-Id: I1b7511c3c55946c87cf381c21b21fe28dd487c06