util: Add 'sudo' to rm WORK_DIR command
Unfortunately Actions uses docker contaienrs to create files on the
system with root permissions. The 'vagrant' user which we login to run
the Actions Runner, can't remove these files. However, 'vagrant' is part
of the sudo group and can therefore use sudo to remove these files.

I don't like this, but it works.

Change-Id: If32a3f47f00ca13d5c8f2f31fc8e40e728bfdca0