arch-riscv: Fix issue when vl=0 in VectorIntMaskMacroConstructor
In VectorIntMaskMacroConstructor, a nop is pushed to microops
when vl=0, bypassing the VectorIntMaskMicro instruction.
This skips the %(copy_old_vd)s section in VectorIntMaskMicroExecute,
which copies old vd values to the result register, vd.
As a result, vd incorrectly stores all zeros.

This PR adjusted the logic in VectorIntMaskMacroConstructor
to ensure the %(copy_old_vd)s section is not skipped when vl = 0,
ensuring correct values in vd.

Change-Id: I37ce386de5238007edb31b7fafbe03bbf6645f06