stdlib: Extend SE Binary Workload std files to accept strings
This patch improves the stdin_file, stdout_file, stderr_file processing in SE Binary Workload.

Earlier, it assumed that Paths would be passed into the function.

Therefore, these parameters could not be defined
in the Resource JSON file as a string to use as is,
since the processing functions like as_posix() would fail.

This is a current pain point of trying to make a Suite JSON
of SPEC Resources, since they use the stdin, stdout and stderr files.

This patch converts the string of the Path in the JSON file into a Path object,
while also retaining the old functionality of allowing Paths to be passed in.

The processing for both remains the same.

Change-Id: I83f281ad4875dfb8db6e088908a8e122299f11bb