python,util: Fix magic number check in decode_inst_dep_trace.py
The decode_inst_dep_trace.py opens the trace file in read mode,
and subsequently reads the magic number from the trace file. Once
the number is read, it is compared against the string 'gem5'
without decoding it first. This causes the comparison to fail.
The fix addresses this by calling the decode() routine on the
output of the read() call.

Github issue: https://github.com/gem5/gem5/issues/543