sim: Rework the Linux Kernel exit events
This patch reworks the Linux Kernel panic and oops events. The code
has been re-factored to provide re-usable events that can be applied
to all ISAs from the base `KernelWorkload` `SimObject`. At the moment
they are installed for the Arm workloads.

This update also provides more configuration options that can be
specified using the new `KernelPanicOopsBehaviour` enum. The options
are applied to the Kernel Workload parameters `on_panic` and `on_oops`
which are available to all subclasses of `KernelWorkload`.

The main rationale for this reworking is to add the option to cleanly
exit the simulation after dumping the Dmesg buffer. Without this
option, the simulation would continue running after a Kernel panic. If
system components (e.g. a system timer) keep the event queue alive,
this causes the simulation to run slowly to the maximum allowed tick.

Change-Id: I2c6dd60f37c542ef66bf6d478a3b97279a2d6dc8
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>