scons: Add an option to reduce memory usage of ld
Linking the gem5 binary consists of linking numerous object files.
By default, ld optimizes for speed over for memory consumption [1].
This leads to the huge memory consumption of gem5 at the linking
stage.

This patch adds an option to add the `--no-keep-memory` flag to ld.
According to the documentation [1], this flag optimizes the memory
usage of ld.

[1] https://ftp.gnu.org/old-gnu/Manuals/ld-2.9.1/html_chapter/ld_2.html#IDX132

Change-Id: I1c41f041dd1463a3210775762dfacb4b8457a80e
Signed-off-by: Hoa Nguyen <hn@hnpl.org>