sim,python: Restore sigint handler in python
Currently, if you try to use ctrl-c while python code is running nothing
happens. This is not ideal. This change enables users to use ctrl-c
while python is running (e.g., when a large disk image is downloading).
To do this, we add a new function that only initializes the sigint
handler and call that just before the event loop begins processing.
Then, when the event loop is done, we reinstall the original handler.

Change-Id: I14490e48d931eb316e8c641217bf8d8ddaa340ed
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>