arch-arm: Handle invalid case for encodeAArch64SysReg
This patch is amending encodeAArch64SysReg so that it covers
the case where there are no arch numbers available for the
misc index passed as an argument.

This could happen if the register ID is a gem5 pseudo register
which is not associated with any architected op1/op2/crn/crm
tuple.

Rather than panicking we return a nullopt.

Change-Id: I7ab70467105ef93c0c78ac4e999c7dc8e5e09925
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>