misc: Make dockerfiles use gem5 source
The gem5-all-min-dependencies dockerfile needs to get the most recent
source with ADD. It doesn't work to use git clone because docker
doesn't see that as new and ignores it.

This commit changes all of the docker images so that it's possible to
use the gem5 source as context. Most of them don't need this, but it
doesn't hurt to add it. This will reduce the chance of a similar mistake
in the future.

Change-Id: I35da631cdd5f48f38c498d3f787bf51b96fbdc62
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>