mem: Make SimpleMemory Backdoor optional
Sometimes a SimpleMemory is placed under a default port as the default
destination of an XBar. Some memory_side_ports are connected, so the
default port might not be continuous. It would be convenient if we could
just assign a universal address range to the SimpleMemory the XBar would
filter which address to be forwarded.
However, since SimpleMemory returns the backdoor for its whole address
range, the CPU would access all address range via the backdoor, even
though some of them were to be distributed to other mem_side_ports.
Thus, it require a device memory with the backdoor disabled.
This CL makes the backdoor optional to fulfill this usecase.

Change-Id: I98c490bed98c26f0372556840669742f876be74d
Signed-off-by: handsomeliu <handsomeliu@google.com>