python: Adding fatal to check for unproxied params
This change adds a fatal statement to check all params for all
SimObjects have been unproxied before C++ object are created.
The fatal statement notifies the user of a mistake that could
possibly lead to a SimObject to not have its params unproxied.
This mistake could be made by adding a child SimObject with a
name that starts with an underscore.

Change-Id: I589fdeb44e8b977a9f43e43a3da8939b6d90113d