arch-riscv: Move alignment check to Physical Memory Attribute(PMA)
In the RISC-V unprivileged spec[1], the misaligned load/store support is
depend on the EEI.

In the RISC-V privileged spec ver1.12[2], the PMA specify wether the
misaligned access is support for each data width and the memory region.

In the [3] of mcause spec, we cloud directly raise misalign exception if
there is no memory region misalignment support. If the part of memory
region support misaligned-access, we need to translate the vaddr to
paddr first then check the paddr later. The page-fault or access-fault
is rose before misalign-fault.

The benefit of moving check_alignment option from ISA option to PMA
option is we can specify the part region of memory support misalign
load/store.

MMU will check alignment with virtual addresss if there is no
misaligned memory region specified. If there are some misaligned memory
region supported, translate address first and check alignment at final.

[1] https://github.com/riscv/riscv-isa-manual/blob/main/src/rv32.adoc#base-instruction-formats
[2] https://github.com/riscv/riscv-isa-manual/blob/main/src/machine.adoc#physical-memory-attributes
[3] https://github.com/riscv/riscv-isa-manual/blob/main/src/machine.adoc#machine-cause-register-mcause

Change-Id: I8e3b1c1ff1910e690bbe01b3de0089765027b3dd