sim-se,cpu-kvm: Fix SE workload setup on KVM CPUs
Running SE workloads on KVM CPUs under the new interface caused
an immediate crash. This commit fixes that by checking if any
of the cores are KVM and if so, configuring the underlying
process accordingly.

See configs/deprecated/se.py:224 to see how this was performed
in the deprecated SE script (which does not have this issue).

Change-Id: I006de46310f96074dc5190646a403b3d280b5791