stdlib,configs,tests: Add gem5 MuliSim (MultiProcessing for gem5)
This allows for multiple gem5 simulations to be spawned from a single
parent gem5 process, as defined in a simgle gem5 configuration. In this
design _all_ the `Simulator`s are defined in the simulation script and
then added to the mutlisim module. For example:

```py
from gem5.simulate.Simulator import Simulator
import gem5.utils.multisim as multisim

simulator1 = Simulator(board=board[0])
simulator2 = Simulator(board=board[2])

multisim.add_simulator(simulator1, id="board-1")
multisim.add_simulator(simulator2, id="board-2")
```

This specifies that two simulations are to be run in parallel in
seperate threads: one specified by `simulator1` and another by
`simulator2`. They are then added to MultiSim via the
`multisim.add_simulator` function. This function takes a simulator and
and id. The user must specify this id and it must be unique to that
simulator. This is used to give each process a unique id and output
directory name. Given this, the id should be a helpful name describing
the simulation being specified.

To run these simulators we use `<gem5 binary> -m gem5.utils.multisim
<script> -p <num_processes>`. Note:multisim is an executable module in gem5. This is the
same module we input into our scripts to add the simulators. This is an
intentionally modular encapsulated design. When the module processes a
script it will schedule multiple gem5 jobs and, dependent
on the number of processes specified, will create child gem5 processes
to processes tjese jobs (jobs are just gem5 simulations in this case).
The `--processes` (`-p`) argument is optional and if not specified the
max number of processes which can be run concurrently will be the number
of available threads on the host system.

The id for each process is used to create a subdirectory inside the
`outputdor` (`m5out`) of that id name. E.g, in the example above the
ID's are `board-0` and `board-1`. Therefore the m5 out directory will
look as follows:

```sh
- m5out
    - board-1
        - stats.txt
        - config.ini
        - config.json
        - terminal.out
    - board-1
        - stats.txt
        - config.ini
        - config.json
        - terminal.out
```

Each simulations output is encapsulated inside the subdirectory of the
id name.

If the multisim configuation script is passed directly to gem5 (like a
traditional gem5 configuraiton script, i.e.: `<gem5 binary> <script>`),
the user may run a single simulation specified in that script by passing
its id as an argument. E.g. `<gem5 binary> <script> board-1` will run
the `board-1` simulation specified in `script`. If no argument is passed
an Exception is raised asking the user to either specify or use the
MultiSim module if multiprocessing is needed.

If the user desires a list of ids of the simulations specified in a
given MultiSim script, they can do so by passing the `--list` (`-l`)
parameter to the config script. I.e., `<gem5 binary> <script> --list`
will list all the IDs for all the simulations specified in`script`.

This change comes with two new example scripts found in
'configs/example/gem5_library/multsim" to demonstrate multisim in both
an SE and FS mode simulation. Tests have been added which run these
scripts as part of gem5' Daily suite of tests.

Notes
=====

* **Bug fixed**: The some of the classic cache hierarchies have been
modified so the Xbar is no longer set with a `__func__` call. This
interfered with MultiProcessing as this structure is not serializable
via Pickle. This was quite bad design anyway so should is a welcome
change.

* **Removal note**: This implementation supercedes the functionality
outlined in 'src/python/gem5/utils/multiprocessing'. As such, this code
has been removed.

Limitations/Things to Fix/Improve
=================================

* Passing a list of Simulator objects means the MultiSim module calls
  `run`. Ergo users can't use `simulator.run(<max_ticks>)`. However,
  this should be available via a `set` function in the simulator. This
  may require some refactoring in some cases.
* Though each Simulator process has its own output directory (a
  subdirectory within m5out, with an ID set by the user unique to that
  Simulator), the stdout and stderr are still output to the terminal,
  not the output directory. This results in:
    1. stdout and stderr data lost and not recorded for these runs.
    2. An incredibly noisy terminal output.
* If a 'readfile' is created the readfile is set to the parent root
  directory, not the individual run. if the readfile is different for
  each simulation, this can cause bugs as the file is overwritten
  constantly.
* Each process uses the same cached resources. If not present many
  threads can end up downloading the same resources at the same time. In
  a similar vein, processes often try to obtain resources at the same
  time causing the a lot of slowdown.
* All available threads are taken. The user cannot specify the number of
  processes.
* There's no real orchastration framework for these processes. running
  many in parallel can be difficult to manage.

Change-Id: Ief5a3b765070c622d1f0de53ebd545c85a3f0eee